+++ apps/manifest-demo/3010-rspack-provider/rspack.config.js
@@ -101,6 +101,9 @@ module.exports = composePlugins(
           },
         },
         dataPrefetch: true,
+        experiments: {
+          externalRuntime: true,
+        },
       }),
     );
     (config.devServer = {