+++ packages/rspack/src/index.ts
@@ -1,4 +1,4 @@
-export { ModuleFederationPlugin } from './ModuleFederationPlugin';
+export { ModuleFederationPlugin, PLUGIN_NAME } from './ModuleFederationPlugin';
 import { container } from '@rspack/core';
 export const ContainerPlugin = container.ContainerPlugin;
 export const ContainerReferencePlugin = container.ContainerReferencePlugin;