+++ packages/bridge/bridge-react/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @module-federation/bridge-react
 
+## 0.8.7
+
+### Patch Changes
+
+- 4102786: feat: support basename passed by remote module props
+- Updated dependencies [835b09c]
+- Updated dependencies [336f3d8]
+- Updated dependencies [4fd33fb]
+  - @module-federation/sdk@0.8.7
+  - @module-federation/bridge-shared@0.8.7
+
+## 0.8.6
+
+### Patch Changes
+
+- 85e7482: fix(bridge-react): change all logs from info to debug in all bridge packages
+  - @module-federation/sdk@0.8.6
+  - @module-federation/bridge-shared@0.8.6
+
 ## 0.8.5
 
 ### Patch Changes