+++ packages/bridge/bridge-shared/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @module-federation/bridge-shared
 
+## 0.8.7
+
+## 0.8.6
+
 ## 0.8.5
 
 ## 0.8.4