+++ packages/managers/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @module-federation/managers
 
+## 0.8.7
+
+### Patch Changes
+
+- Updated dependencies [835b09c]
+- Updated dependencies [336f3d8]
+- Updated dependencies [4fd33fb]
+  - @module-federation/sdk@0.8.7
+
+## 0.8.6
+
+### Patch Changes
+
+- @module-federation/sdk@0.8.6
+
 ## 0.8.5
 
 ### Patch Changes