+++ packages/rspack/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @module-federation/rspack
 
+## 0.8.7
+
+### Patch Changes
+
+- Updated dependencies [835b09c]
+- Updated dependencies [f573ad0]
+- Updated dependencies [336f3d8]
+- Updated dependencies [4fd33fb]
+  - @module-federation/sdk@0.8.7
+  - @module-federation/inject-external-runtime-core-plugin@0.8.7
+  - @module-federation/bridge-react-webpack-plugin@0.8.7
+  - @module-federation/dts-plugin@0.8.7
+  - @module-federation/managers@0.8.7
+  - @module-federation/manifest@0.8.7
+  - @module-federation/runtime-tools@0.8.7
+
+## 0.8.6
+
+### Patch Changes
+
+- Updated dependencies [c90bba2]
+- Updated dependencies [4a2cf82]
+  - @module-federation/dts-plugin@0.8.6
+  - @module-federation/manifest@0.8.6
+  - @module-federation/sdk@0.8.6
+  - @module-federation/runtime-tools@0.8.6
+  - @module-federation/managers@0.8.6
+  - @module-federation/bridge-react-webpack-plugin@0.8.6
+
 ## 0.8.5
 
 ### Patch Changes